[build] fix mandir installation
When building and installing (build for Debian package at least) it
creates a directory /man8 that should not exist. The reason for this
is that src/Makefile use MANDIR while this directory is set in
man/Makefile.

At the same time man/Makefile have problem to create the man8 directory,
because INSTALL is not defined.

Fix both.

[kir@openvz.org: remove MANDIR= from make install line in .spec]

https://bugzilla.openvz.org/2710